Adding precommit hook (#34)
