+++ package.json
@@ -50,9 +50,13 @@
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-unused-imports": "^2.0.0",
     "git-url-parse": "^11.6.0",
+    "pre-commit": "^1.2.2",
     "prettier": "^2.5.1",
     "simple-git": "^2.48.0",
     "typescript": "^4.5.4",
     "yaml": "^1.10.2"
-  }
+  },
+  "pre-commit": [
+    "lint"
+  ]
 }
