--- src/components/Footer/Footer.module.css
@@ -1,13 +1,0 @@
-.footer {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	left: 0;
-	bottom: 0;
-	width: 100%;
-    height: 50px;
-    padding: 10px;
-	background-color: black;
-	color: white;
-	text-align: center;
-}
